python/requirements.txt
google-auth==1.5.1
requests==2.20.0
